Fixing issues related to PR #178
