+++ gpu_ocean/SWESimulators/KP07.py
@@ -85,14 +85,6 @@ class KP07(Simulator.Simulator):
         flux_slope_eps: Used for desingularization with dry cells
         """
        
-        ## After changing from (h, B) to (eta, H), several of the simulator settings used are wrong. This check will help detect that.
-        #if ( np.sum(eta0 - H[:-1, :-1] > 0) > nx):
-        #    assert(False), "It seems you are using water depth/elevation h and bottom topography B, while you should use water level eta and equillibrium depth H."
-
-        
-
-        
-        
         
         ghost_cells_x = 2
         ghost_cells_y = 2
